refactor: 패키지 구조 변경
report 도메인 폴더를 따로 생성하여, 정리해두었습니다.
